Store payment amount as a bigint

+++ views/transactions_payment.sql
@@ -10,7 +10,7 @@ create view transactions_payment as
 			, fields::json->>'nonce' as nonce
 			, fields::json->>'payee' as payee
 			, fields::json->>'payer' as payer
-			, fields::json->>'amount' as amount    
+			, fields::json->>'amount' as amount::bigint
 	from transactions 
 	where type IN ('payment_v1', 'payment_v2');
